ENH: make positionfixes import circle free
