+++ trackintel/geogr/distances.py
@@ -11,6 +11,9 @@ import similaritymeasures
 from scipy.spatial.distance import cdist
 from sklearn.metrics import pairwise_distances
 
+from trackintel import Positionfixes
+from trackintel.model.util import doc
+
 
 def point_haversine_dist(lon_1, lat_1, lon_2, lat_2, r=6371000, float_flag=False):
     """
@@ -80,54 +83,8 @@ def point_haversine_dist(lon_1, lat_1, lon_2, lat_2, r=6371000, float_flag=False
     return r * np.arccos(cos_lat_d - cos_lat1 * cos_lat2 * (1 - cos_lon_d))
 
 
+@doc(Positionfixes.calculate_distance_matrix, first_arg="\nX : GeoDataFrame (as trackintel staypoints or triplegs)\n")
 def calculate_distance_matrix(X, Y=None, dist_metric="haversine", n_jobs=0, **kwds):
-    """
-    Calculate a distance matrix based on a specific distance metric.
-
-    If only X is given, the pair-wise distances between all elements in X are calculated. If X and Y are given, the
-    distances between all combinations of X and Y are calculated. Distances between elements of X and X, and distances
-    between elements of Y and Y are not calculated.
-
-    Parameters
-    ----------
-    X : GeoDataFrame (as trackintel staypoints or triplegs)
-
-    Y : GeoDataFrame (as trackintel staypoints or triplegs), optional
-
-    dist_metric: {'haversine', 'euclidean', 'dtw', 'frechet'}
-        The distance metric to be used for calculating the matrix.
-
-        For staypoints, common choice is 'haversine' or 'euclidean'. This function wraps around
-        the ``pairwise_distance`` function from scikit-learn if only `X` is given and wraps around the
-        ``scipy.spatial.distance.cdist`` function if X and Y are given.
-        Therefore the following metrics are also accepted:
-
-        via ``scikit-learn``: `[‘cityblock’, ‘cosine’, ‘euclidean’, ‘l1’, ‘l2’, ‘manhattan’]`
-
-        via ``scipy.spatial.distance``: `[‘braycurtis’, ‘canberra’, ‘chebyshev’, ‘correlation’, ‘dice’, ‘hamming’, ‘jaccard’,
-        ‘kulsinski’, ‘mahalanobis’, ‘minkowski’, ‘rogerstanimoto’, ‘russellrao’, ‘seuclidean’, ‘sokalmichener’,
-        ‘sokalsneath’, ‘sqeuclidean’, ‘yule’]`
-
-        For triplegs, common choice is 'dtw' or 'frechet'. This function uses the implementation
-        from similaritymeasures.
-
-    n_jobs: int
-        Number of cores to use: 'dtw', 'frechet' and all distance metrics from `pairwise_distance` (only available
-        if only X is given) are parallelized.
-
-    **kwds:
-        optional keywords passed to the distance functions.
-
-    Returns
-    -------
-    D: np.array
-        matrix of shape (len(X), len(X)) or of shape (len(X), len(Y)) if Y is provided.
-
-    Examples
-    --------
-    >>> calculate_distance_matrix(staypoints, dist_metric="haversine")
-    >>> calculate_distance_matrix(triplegs_1, triplegs_2, dist_metric="dtw")
-    """
     geom_type = X.geometry.iat[0].geom_type
     if Y is None:
         Y = X
@@ -325,25 +282,15 @@ def calculate_haversine_length(gdf):
     return np.bincount((index[:-1])[no_mix], weights=dist[no_mix])
 
 
+@doc(
+    Positionfixes.get_speed,
+    first_arg="""
+Parameters
+----------
+positionfixes : GeoDataFrame (as trackintel positionfixes)
+""",
+)
 def get_speed_positionfixes(positionfixes):
-    """
-    Compute speed per positionfix (in m/s)
-
-    Parameters
-    ----------
-    positionfixes : GeoDataFrame (as trackintel positionfixes)
-        The positionfixes have to follow the standard definition for positionfixes DataFrames.
-
-    Returns
-    -------
-    pfs: GeoDataFrame (as trackintel positionfixes)
-        The original positionfixes with a new column ``[`speed`]``. The speed is given in m/s
-
-    Notes
-    -----
-    The speed at one positionfix is computed from the distance and time since the previous positionfix. For the first
-    positionfix, the speed is set to the same value as for the second one.
-    """
     pfs = positionfixes.copy()
     is_planar_crs = check_gdf_planar(pfs)
